Separate Task and Handler classes
Tasks and Handlers have typically been managed in the same way
but there is an important difference, in checking whether something
actually needs to be run.

Handlers are only ever run when they are notified by a task (or another
handler). So they don't need `creates` or `removes` or `when` to detect
this.

This allows standards checks for idempotency to only check tasks