Migrate from flake8 to pycodestlye
Flake8 was renamed to pycodestyle long time ago and for compatibility
the old package would install pycodetyle but only a specific version
which may cause conflicts with other tools

This change adopts pycodestyle so we would avoid dependency conflicts.

At the same time, we introduced a 'pip check' execution in CI in
order to spot future conflicts that could be introduced by
the installation of ansible-review.